requirements.txt
asgiref==3.7.2
Django==4.2.4
docx2pdf==0.1.8
fire==0.5.0
fonttools==4.42.0
lxml==4.9.3
numpy==1.25.2
opencv-python==4.8.0.76
pdf2docx==0.5.6
pdfrw==0.4
Pillow==10.0.0
PyMuPDF==1.22.5
pypdf==3.15.1
PyPDF2==3.0.1
python-docx==0.8.11
reportlab==4.0.4
six==1.16.0
sqlparse==0.4.4
termcolor==2.3.0
tqdm==4.66.1
